Test_codes/requirements.txt
alabaster==0.7.16
asciitree==0.3.3
astroplan==0.9.1
astropy==6.0.0
astropy-healpix==1.0.2
astropy-iers-data==0.2024.2.5.0.30.52
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrs==23.2.0
avro-python3==1.10.2
awkde @ git+https://github.com/titodalcanton/awkde.git@cd782902efad3823a4a09274bd98dc9d597285c6
Babel==2.14.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
contourpy==1.2.0
corner==2.2.2
cpnest==0.11.5
cryptography==42.0.2
cycler==0.12.1
Cython==3.0.8
dask==2024.2.0
dateparser==1.2.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docutils==0.20.1
dqsegdb2==1.2.1
dynesty==2.1.3
emcee==2.2.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
epsie==1.0.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fasteners==0.19
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
glasflow==0.3.0
gwdatafind==1.2.0
gwosc==0.7.1
gwpy==3.0.8
h5py==3.10.0
healpy==1.16.6
idna==3.6
igwn-auth-utils==1.1.0
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1707326244187/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
joblib==1.3.2
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1704727023078/work
kiwisolver==1.4.5
lalsuite==7.21
ligo-gracedb==2.12.0
ligo-segments==1.4.0
ligo.skymap==1.1.2
ligotimegps==2.0.1
llvmlite==0.42.0
locket==1.0.0
lscsoft-glue==3.0.2
Mako==1.3.2
MarkupSafe==2.0.1
matplotlib==3.8.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mpld3==0.5.10
mpmath==1.3.0
nessai==0.11.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
numba==0.59.0
numcodecs==0.12.1
numpy==1.26.4
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.2.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
partd==1.4.1
patsy==0.5.6
pegasus-wms.api==5.0.6
pegasus-wms.common==5.0.6
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptemcee @ https://github.com/willvousden/ptemcee/archive/master.tar.gz#sha256=b7e34c7a2061d5af54f2f8ca1dce45308abd2ad21117b734832edff63c0f96d8
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pybind11==2.11.1
PyCBC @ file:///home/arush/GW_Test_1/pycbc
pycparser==2.21
pyerfa==2.0.1.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
PyJWT==2.8.0
pykerr==0.1.0
pymultinest==2.12
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyRXP==3.0.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-ligo-lw==1.8.3
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
regex==2023.12.25
reproject==0.13.0
requests==2.31.0
safe-netrc==1.0.1
scikit-learn==1.4.0
scipy==1.12.0
scitokens==1.8.1
seaborn==0.13.2
setproctitle==1.3.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-carousel==1.2.0
sphinx-rtd-theme==2.0.0
sphinx_design==0.5.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-programoutput==0.17
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
statsmodels==0.14.1
stumpy==1.12.0
sympy==1.12
threadpoolctl==3.2.0
toolz==0.12.1
torch==2.2.0+cpu
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
tsfresh==0.20.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1702176139754/work
tzdata==2023.4
tzlocal==5.2
ultranest==3.6.4
urllib3==2.2.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
zarr==2.16.1
zipp==3.17.0
